+++ CrystalPhysics/CrystalPhysics/MVPSurfaceBuilder.h
@@ -17,7 +17,7 @@ class MVPSurfaceBuilder
 public:
 	void build(const std::vector<MVPVolumeParticle*>& particles, const int res, const double threshold);
 
-	void buildVolumes(const std::vector<Math::Vector3dd>& volumeParticles, const double radius, const int res);
+	void buildVolumes(const std::vector<Math::Vector3dd>& volumeParticles, const float radius, const int res);
 
 	void buildMasses(const std::vector<Math::Vector3dd>& massParticles, const double radius);
